+++ common/src/main/java/com/bards/client/armor/CustomArmorRenderer.java
@@ -1,30 +1,29 @@
 package com.bards.client.armor;
 
-import mod.azure.azurelibarmor.common.render.armor.AzArmorRenderer;
-import mod.azure.azurelibarmor.common.render.armor.AzArmorRendererConfig;
 import net.minecraft.util.Identifier;
+import net.rpg_foundation.armor_api.client.GeoArmorRenderer;
 
 import static com.bards.BardsMod.MOD_ID;
 
-public class CustomArmorRenderer extends AzArmorRenderer {
+public final class CustomArmorRenderer {
+    private CustomArmorRenderer() { }
 
-    public static CustomArmorRenderer entertainer_armor() {
-        return new CustomArmorRenderer("entertainer_armor", "entertainer_armor");
+    public static GeoArmorRenderer entertainer_armor() {
+        return make("entertainer_armor", "entertainer_armor");
     }
-    public static CustomArmorRenderer troubadour_armor() {
-        return new CustomArmorRenderer("troubadour_armor", "troubadour_armor");
+    public static GeoArmorRenderer troubadour_armor() {
+        return make("troubadour_armor", "troubadour_armor");
     }
-    public static CustomArmorRenderer netherite_troubadour_armor() {
-        return new CustomArmorRenderer("troubadour_armor", "netherite_troubadour_armor");
+    public static GeoArmorRenderer netherite_troubadour_armor() {
+        return make("troubadour_armor", "netherite_troubadour_armor");
     }
-    public static CustomArmorRenderer storyteller_armor() {
-        return new CustomArmorRenderer("storyteller_armor", "storyteller_armor");
+    public static GeoArmorRenderer storyteller_armor() {
+        return make("storyteller_armor", "storyteller_armor");
     }
 
-    public CustomArmorRenderer(String modelName, String textureName) {
-        super(AzArmorRendererConfig.builder(
+    private static GeoArmorRenderer make(String modelName, String textureName) {
+        return GeoArmorRenderer.of(
                 Identifier.of(MOD_ID, "geo/" + modelName + ".geo.json"),
-                Identifier.of(MOD_ID, "textures/armor/" + textureName + ".png")
-        ).build());
+                Identifier.of(MOD_ID, "textures/armor/" + textureName + ".png"));
     }
 }