+++ common/src/main/java/com/bards/config/TweaksConfig.java
@@ -5,4 +5,5 @@ public class TweaksConfig {
     public float armies_paeon_impact_multiplier = 0.035F;
     public float discordant_selfdamage_chance_per_stack = 0.02F;
     public float discordant_selfdamage_multiplier_per_stack= 0.1F;
+    public float luthier_song_volume = 1.0F;
 }